refactor: rename models
