+++ src/domain/models/auth-model.ts
