+++ src/domain/models/index.ts
@@ -1,4 +1,4 @@
-export * from './user'
-export * from './link'
-export * from './auth'
-export * from './social'
+export * from './user-model'
+export * from './link-model'
+export * from './auth-model'
+export * from './social-model'
