+++ src/domain/models/link-model.ts
