+++ src/domain/models/social-model.ts
